# opt: use switches instead of function tables