--- a/pkg/sql/opt/optgen/cmd/optgen/exprs_gen.go
+++ b/pkg/sql/opt/optgen/cmd/optgen/exprs_gen.go
@@ -39,7 +39,7 @@ func (g *exprsGen) generate(compiled *lang.CompiledExpr, w io.Writer) {
 	fmt.Fprintf(g.w, ")\n\n")
 
 	g.genChildCountLookup()
-	g.genChildGroupLookup()
+	g.genChildGroup()
 	g.genPrivateFieldLookup()
 	g.genTagLookup()
 	g.genIsTag()
@@ -94,23 +94,20 @@ func (g *exprsGen) genChildCountLookup() {
 	fmt.Fprintf(g.w, "}\n\n")
 }
 
-// genChildGroupLookup generates a lookup table used to implement the ExprView
-// ChildGroup method for each different kind of memo expression.
-func (g *exprsGen) genChildGroupLookup() {
-	fmt.Fprintf(g.w, "type childGroupLookupFunc func(ev ExprView, n int) opt.GroupID\n")
-
-	fmt.Fprintf(g.w, "var childGroupLookup = [...]childGroupLookupFunc{\n")
-	fmt.Fprintf(g.w, "  // UnknownOp\n")
-	fmt.Fprintf(g.w, "  func(ev ExprView, n int) opt.GroupID {\n")
-	fmt.Fprintf(g.w, "    panic(\"op type not initialized\")\n")
-	fmt.Fprintf(g.w, "  },\n\n")
+// genChildGroup generates the ExprView.ChildGroup method.
+func (g *exprsGen) genChildGroup() {
+	fmt.Fprintf(g.w, "// ChildGroup returns the memo group containing the nth child of this parent\n")
+	fmt.Fprintf(g.w, "// expression.\n")
+	fmt.Fprintf(g.w, "func (ev ExprView) ChildGroup(n int) opt.GroupID {\n")
+	fmt.Fprintf(g.w, "  switch ev.op {\n")
+	fmt.Fprintf(g.w, "  case opt.UnknownOp:\n")
+	fmt.Fprintf(g.w, "    panic(\"opt type not initialized\")\n\n")
 
 	for _, define := range g.compiled.Defines {
 		exprType := fmt.Sprintf("%sExpr", unTitle(string(define.Name)))
 		varName := exprType
 
-		fmt.Fprintf(g.w, "  // %sOp\n", define.Name)
-		fmt.Fprintf(g.w, "  func(ev ExprView, n int) opt.GroupID {\n")
+		fmt.Fprintf(g.w, "  case opt.%sOp:\n", define.Name)
 
 		count := len(define.Fields)
 		if privateField(define) != nil {
@@ -118,8 +115,7 @@ func (g *exprsGen) genChildGroupLookup() {
 		}
 
 		if count == 0 {
-			fmt.Fprintf(g.w, "    panic(\"child index out of range\")\n")
-			fmt.Fprintf(g.w, "  },\n\n")
+			fmt.Fprintf(g.w, "    panic(\"child index out of range\")\n\n")
 			continue
 		}
 
@@ -129,8 +125,7 @@ func (g *exprsGen) genChildGroupLookup() {
 			fmt.Fprintf(g.w, "      return ev.loc.group\n")
 			fmt.Fprintf(g.w, "    }\n\n")
 
-			fmt.Fprintf(g.w, "    panic(\"child index out of range\")\n")
-			fmt.Fprintf(g.w, "  },\n\n")
+			fmt.Fprintf(g.w, "    panic(\"child index out of range\")\n\n")
 			continue
 		}
 
@@ -163,9 +158,12 @@ func (g *exprsGen) genChildGroupLookup() {
 			fmt.Fprintf(g.w, "      panic(\"child index out of range\")\n")
 			fmt.Fprintf(g.w, "    }\n")
 		}
-
-		fmt.Fprintf(g.w, "  },\n\n")
+		fmt.Fprintf(g.w, "\n")
 	}
+
+	fmt.Fprintf(g.w, "  default:\n")
+	fmt.Fprintf(g.w, "      panic(\"invalid op\")\n")
+	fmt.Fprintf(g.w, "  }\n\n")
 
 	fmt.Fprintf(g.w, "}\n\n")
 }