--- a/pkg/sql/opt/optgen/cmd/optgen/testdata/exprs
+++ b/pkg/sql/opt/optgen/cmd/optgen/testdata/exprs
@@ -32,16 +32,14 @@ var childCountLookup = [...]childCountLookupFunc{
 	},
 }
 
-type childGroupLookupFunc func(ev ExprView, n int) opt.GroupID
+// ChildGroup returns the memo group containing the nth child of this parent
+// expression.
+func (ev ExprView) ChildGroup(n int) opt.GroupID {
+	switch ev.op {
+	case opt.UnknownOp:
+		panic("opt type not initialized")
 
-var childGroupLookup = [...]childGroupLookupFunc{
-	// UnknownOp
-	func(ev ExprView, n int) opt.GroupID {
-		panic("op type not initialized")
-	},
-
-	// FuncCallOp
-	func(ev ExprView, n int) opt.GroupID {
+	case opt.FuncCallOp:
 		funcCallExpr := (*funcCallExpr)(ev.mem.lookupExpr(ev.loc))
 
 		switch n {
@@ -51,7 +49,11 @@ var childGroupLookup = [...]childGroupLookupFunc{
 			list := ev.mem.lookupList(funcCallExpr.args())
 			return list[n-1]
 		}
-	},
+
+	default:
+		panic("invalid op")
+	}
+
 }
 
 type privateLookupFunc func(ev ExprView) opt.PrivateID
@@ -132,22 +134,24 @@ var childCountLookup = [...]childCountLookupFunc{
 	},
 }
 
-type childGroupLookupFunc func(ev ExprView, n int) opt.GroupID
+// ChildGroup returns the memo group containing the nth child of this parent
+// expression.
+func (ev ExprView) ChildGroup(n int) opt.GroupID {
+	switch ev.op {
+	case opt.UnknownOp:
+		panic("opt type not initialized")
 
-var childGroupLookup = [...]childGroupLookupFunc{
-	// UnknownOp
-	func(ev ExprView, n int) opt.GroupID {
-		panic("op type not initialized")
-	},
-
-	// SortOp
-	func(ev ExprView, n int) opt.GroupID {
+	case opt.SortOp:
 		if n == 0 {
 			return ev.loc.group
 		}
 
 		panic("child index out of range")
-	},
+
+	default:
+		panic("invalid op")
+	}
+
 }
 
 type privateLookupFunc func(ev ExprView) opt.PrivateID