#### opt: add ExprView benchmark

Benchmark the traversal of an ExprView.

Release note: None

#### opt: use a switch instead of function table for ChildGroup

```
name          old time/op    new time/op    delta
ExprView/0-8     430ns ± 1%     395ns ± 0%  -8.13%  (p=0.000 n=10+8)
```

Release note: None

#### opt: use a switch instead of function table for ChildCount

```
name          old time/op    new time/op    delta
ExprView/0-8     397ns ± 2%     342ns ± 1%  -13.84%  (p=0.000 n=10+10)
```

Release note: None