@RaduBerinde 2018-03-02
I plan to do the same with `private`, but in a separate PR (the benchmark will need to be modified).