@andy-kimball 2018-03-02
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>
Interesting. I'm a bit worried that once we add all the physical operators that this approach will slow down, but I guess we worry about it when it happens...

---

Review status: 0 of 6 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23316#-:-L6bvjNoOIN0U9X15twL:bsc4fu5)*
<!-- Sent from Reviewable.io -->