@RaduBerinde 2018-03-02
Yeah, I'm not crazy about the generated code (basically a ton of conditional jumps). I will experiment with something better, at least for `ChildCount` (we might be able to generate a table of values storing the number of children plus some flags).